+++ dev/dev_history_cran.R
@@ -26,8 +26,10 @@ testthat::test_dir("tests/testthat/")
 testthat::test_file("tests/testthat/test-inflate-part1.R")
 testthat::test_file("tests/testthat/test-inflate-part2.R")
 testthat::test_file("tests/testthat/test-inflate_all.R")
+testthat::test_file("tests/testthat/test-inflate_all_utils.R")
 testthat::test_file("tests/testthat/test-add_flat_template.R")
 testthat::test_file("tests/testthat/test-skeleton.R")
+testthat::test_file("tests/testthat/test-register_config_file.R")
 Sys.setenv("NOT_CRAN" = "false")
 
 # Run line by line